// src/TadbirNG/SPPC.Tadbir.Web/ClientApp/src/app/config/models/settingBrief.ts
import { IEntity } from "@sppc/shared/models";

// ------------------------------------------------------------------------------
// <auto-generated>
//     This code was generated by a tool.
//     Runtime Version: 1.1.340
//     Template Version: 1.0
//     Generation Date: 07/11/2018 04:05:57 ب.ظ
//  
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.
// </auto-generated>
// ------------------------------------------------------------------------------

//import { Object } from "./object";

export interface SettingBrief extends IEntity{
    id: number;
    title: string;
    description?: string;
    modelType: string;
    values: Object;
    defaultValues: Object;
}